Add some functionnality and fix

+++ FrontReactJs/src/features/users/UserList.tsx
@@ -1,5 +1,8 @@
 import { useEffect, useState } from 'react';
+import { useRouter } from 'next/router';
+import { useSession } from 'next-auth/react';
 import { useDispatch, useSelector } from 'react-redux';
+
 import { Box, IconButton } from '@mui/material';
 import { DataGrid, GridColDef } from '@mui/x-data-grid';
 
@@ -11,7 +14,7 @@ import { fetchUsersAsync } from './UserSlice';
 import Loading from '@/components/common/Loading';
 import { Dialog } from '@/components/common/dialog';
 import { User } from '@/models/user/user';
-import { useRouter } from 'next/router';
+import { decryptPayload } from '@/utils/encryptUtils';
 
 const UserList: React.FC = () => {
   const dispatch = useDispatch();
@@ -19,6 +22,10 @@ const UserList: React.FC = () => {
   const status = useSelector((state: RootState) => state.users.status);
   const error = useSelector((state: RootState) => state.users.error);
   const router = useRouter();
+
+  const { data: session } = useSession();
+  let token: string = '';
+  let right: string = '';
 
   const [openDialog, setOpenDialog] = useState(false);
   const [selectedUser, setSelectedUser] = useState<User | null>(null);
@@ -40,6 +47,31 @@ const UserList: React.FC = () => {
     return <div>Error: {error}</div>;
   }
 
+  // Check if the session is available and contains the encrypted session
+  if (session && session.user && session.user.encryptedSession) {
+    const encryptedSession = session.user.encryptedSession; // Retrieve encrypted session data
+
+    // Extract encrypted data and IV (if needed)
+    const { encryptedData, iv } = encryptedSession;
+
+    // Decrypt the session data
+    try {
+      const decryptedSessionData = decryptPayload(encryptedData, iv);
+
+      // Cast the decrypted session data to the expected structure
+      const { token: decryptedToken, right: decryptedRight } =
+        decryptedSessionData as {
+          token: string;
+          right: string;
+        };
+
+      token = decryptedToken;
+      right = decryptedRight;
+    } catch (error) {
+      console.error('Failed to decrypt session data:', error);
+    }
+  }
+
   const handleEdit = (user: User) => {
     setSelectedUser(user);
     setDialogTitle('Edit User');
@@ -54,7 +86,9 @@ const UserList: React.FC = () => {
   const handleDelete = (user: User) => {
     setSelectedUser(user);
     setDialogTitle('Delete User');
-    setDialogContent(`Are you sure you want to delete user: ${user.userLogin}?`);
+    setDialogContent(
+      `Are you sure you want to delete user: ${user.userLogin}?`,
+    );
     setDialogAction(() => () => {
       // dispatch(deleteUser(user.userId));
       handleCloseDialog();
@@ -73,41 +107,52 @@ const UserList: React.FC = () => {
   }));
 
   const columns: GridColDef[] = [
-    { field: 'userId', headerName: 'ID' },
+    ...(right !== null && (right === 'Admin' || right === 'Super Admin')
+      ? [{ field: 'userId', headerName: 'ID' }]
+      : []),
     { field: 'userFirstname', headerName: 'Prénom' },
     { field: 'userLastname', headerName: 'Nom' },
     { field: 'userLogin', headerName: 'Login' },
     { field: 'userEmail', headerName: 'E-mail', width: 200 },
-    { field: 'userRight', headerName: 'Droit' },
+    { field: 'userRight', headerName: 'Droit', width: 130 },
     {
       field: 'actions',
       headerName: 'Actions',
       width: 150,
       sortable: false,
-      renderCell: (params) => (
-        <>
-          <IconButton
-            color="primary"
-            aria-label="edit user"
-            onClick={() => handleEdit(params.row)}
-          >
-            <EditIcon />
-          </IconButton>
-          <IconButton
-            color="secondary"
-            aria-label="delete user"
-            onClick={() => handleDelete(params.row)}
-          >
-            <DeleteIcon />
-          </IconButton>
-        </>
-      ),
+      renderCell: params => {
+        // Only show action buttons based on logged-in user's rights
+        const userRight = params.row.userRight;
+
+        // Admin cannot modify Super Admin users
+        if (right === 'Admin' && userRight === 'Super Admin') {
+          return null;
+        }
+        return (
+          <>
+            <IconButton
+              color='primary'
+              aria-label='edit user'
+              onClick={() => handleEdit(params.row)}
+            >
+              <EditIcon />
+            </IconButton>
+            <IconButton
+              color='secondary'
+              aria-label='delete user'
+              onClick={() => handleDelete(params.row)}
+            >
+              <DeleteIcon />
+            </IconButton>
+          </>
+        );
+      },
     },
   ];
 
   return (
     <Box>
-      <DataGrid 
+      <DataGrid
         rows={rows}
         columns={columns}
         checkboxSelection
